splitting admins from spikers, modifying spiker add and edit
